requirements.txt
django==2.2.7
django-model-utils==3.2.0